requirements.txt
pandas==1.1.1
tensorboard==2.2.0
torch==1.6.0
transformers==3.1.0
pytorch-lightning==1.4.5
pytorch-nlp==0.5.0
matplotlib
seaborn
sklearn
jupyter